+++ CHANGELOG.md
@@ -47,3 +47,4 @@
 - 2023-08-07: Async deals, mod, qualifier, inbox, +2
 - 2023-08-08: Impl smtp, mod, mod, mod, +6 query
 - 2023-08-09: Retry CHANGELOG, qualifier, architecture, deals, +2
+- 2023-08-10: Bench inbox, mod, gp, mod, +2